feat: add ShiftReportController with start, end, current progress, and cashier report endpoints

+++ src/main/java/com/tap/payload/dto/ShiftReportDTO.java
@@ -28,7 +28,7 @@ public class ShiftReportDTO {
     private Double totalSales;
     private Double totalRefunds;
     private Double netSale;
-    private Double totalOrders;
+    private int totalOrders;
 
 
     private UserDto cashier;
